Implement DPMS creen on/off control to close #26
Bandshell will now always attempt to enforce the current screen state
over DPMS, with debug information available in /player_status if for
some reason it is not available. The state is enforced periodically
(5 seconds) based on the most recent rules (currently updated every
minute).

The screen is currently "always on" at startup and until the first
update, or until authenticated (for new screens).