Fix a bandshell-timer crash on password change
The symptom that I was seeing was that after entering a new system
password as prompted by the web UI, bandshell-timer would crash,
causing bandshell to stop attempting to control the screen. The
error in the bandshell-timer log was an uninitialized constant
Bandshell::Passwords. I'm not sure how prevalent this issue is.

If you suspect this may be an issue in an existing deployment, check
/var/log/bandshell-timer.rb.output on the player for such a message.