Retrieve timezone and use it to determine on/off times.
Admins can set the screen's timezone in Concerto, and the latest
concerto-hardware plugin will transmit that information to the screen
via the API. The on/off times set in concerto-hardware are read as local
time to the time zone set in the Screen.

The only thing from #45 that this does not resolve is setting the system
time.

Screen name is also now saved to aid in debugging via the status page.